Refactored data passing services into one big service
